Accept PR recommendations
